Restructure tests back to per-topic files
The previous sweep collapsed every topic into a single test/lib/shared.ts,
losing the per-file granularity that the original master/test/*.test.ts
layout offered. This commit decomposes shared.ts into topic-specific
modules under test/lib/ (binary, encoding, synopsis, text, content-length),
each exporting a single run<Topic>Tests(express) function. Common helpers
(ExpressFn type, toHEX, rawRequest) live in test/lib/util.ts.

The two entry points (express4.test.ts / express5.test.ts) import all
five topic modules and run them as describe blocks, preserving the v4 / v5
matrix established by the npm: aliases.

Type-only imports use `import type` to stay compatible with Node's
strip-only TypeScript mode.

Also runs fixpack against package.json (already clean).
