+++ test/express4.test.ts
@@ -1,18 +1,16 @@
 // express4.test.ts — verifies Express 4 behavior.
 //
-// Loads `npm:express@^4` via the `express4` alias and runs every shared
+// Loads `npm:express@^4` via the `express4` alias and runs every per-topic
 // test against v4.
 
 import {describe} from "node:test";
 import express from "express4";
 
-import {
-    runBinaryTests,
-    runContentLengthTests,
-    runEncodingTests,
-    runSynopsisTests,
-    runTextTests,
-} from "./lib/shared.ts";
+import {runBinaryTests} from "./lib/binary.ts";
+import {runContentLengthTests} from "./lib/content-length.ts";
+import {runEncodingTests} from "./lib/encoding.ts";
+import {runSynopsisTests} from "./lib/synopsis.ts";
+import {runTextTests} from "./lib/text.ts";
 
 describe("express4: binary", () => runBinaryTests(express as any));
 describe("express4: encoding", () => runEncodingTests(express as any));
